+++ src/main/java/cz/uhk/pro2/tulipani/domain/repository/TodolistUserRepository.java
@@ -4,12 +4,12 @@ import cz.uhk.pro2.tulipani.domain.entity.TodolistUser;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TodolistUserRepository extends JpaRepository<TodolistUser, Long> {
+public interface TodolistUserRepository extends JpaRepository<TodolistUser, Integer> {
 
-	List<TodolistUser> findByUserId(Long userId);
+	List<TodolistUser> findByUserId(Integer userId);
 
-	boolean existsByTodolistIdAndUserId(Long todolistId, Long userId);
+	boolean existsByTodolistIdAndUserId(Integer todolistId, Integer userId);
 
-	java.util.Optional<TodolistUser> findByTodolistIdAndUserId(Long todolistId, Long userId);
+	java.util.Optional<TodolistUser> findByTodolistIdAndUserId(Integer todolistId, Integer userId);
 
 }
