+++ src/main/java/cz/uhk/pro2/tulipani/web/dto/TodolistResponse.java
@@ -1,7 +1,7 @@
 package cz.uhk.pro2.tulipani.web.dto;
 
 public record TodolistResponse(
-        Long todolistId,
+        Integer todolistId,
         String name,
         String listType) {
 }
